kernelshark: Fix compiler error in KsSession.cpp on Raspberry Pi 3
Fix compiler error at KsSession.cpp:457:30 on Raspberry Pi 3

Fix by changing the type of local variable 'pos' from uint64_t to
size_t in KsSession::loadDualMarker().

KsSession.cpp:457:30: error: no matching function for call to
‘KsSession::_getMarker(const char [6], uint64_t*)’
  if (_getMarker("markA", &pos)) {
                              ^
In file included from KsSession.cpp:14:0:
KsSession.hpp:97:7: note: candidate:
 bool KsSession::_getMarker(const char*, size_t*)
  bool _getMarker(const char* name, size_t *pos);
       ^~~~~~~~~~
KsSession.hpp:97:7: note:   no known conversion for
argument 2 from ‘uint64_t* {aka long long unsigned int*}’
to ‘size_t* {aka unsigned int*}’

Link: http://lore.kernel.org/linux-trace-devel/20190612035902.16366-1-amikhak@wirelessfabric.com

Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=203873

+++ kernel-shark/src/KsSession.cpp
@@ -450,7 +450,7 @@ void KsSession::saveDualMarker(KsDualMarkerSM *dm)
  */
 void KsSession::loadDualMarker(KsDualMarkerSM *dm, KsTraceGraph *graphs)
 {
-	uint64_t pos;
+	size_t pos;
 
 	dm->reset();
 	dm->setState(DualMarkerState::A);
